+++ graphics/sdl2/src/window/Window.hpp
@@ -141,13 +141,13 @@ public:
      * @brief Convert a position in pixels to a position in tiles
      * @return Converted position
      */
-    Vector2i pixelsToTiles(const Vector2i &position) const;
+    Vector2f pixelsToTiles(const Vector2i &position) const;
 
     /**
      * @brief Convert a position in tiles to a position in pixels
      * @return Converted position
      */
-    Vector2i tilesToPixels(const Vector2i &position) const;
+    Vector2i tilesToPixels(const Vector2f &position) const;
 
     /**
      * @brief Convert a position in tiles to a position in pixels
